~ restructure bsd, parser, generator, dictionary manager and reader
