+++ milo-examples/client-examples/pom.xml
@@ -54,6 +54,11 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.milo</groupId>
+            <artifactId>dictionary-reader</artifactId>
+            <version>0.4.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.milo</groupId>
             <artifactId>server-examples</artifactId>
             <version>0.4.0-SNAPSHOT</version>
         </dependency>
