+++ opc-ua-sdk/sdk-client/src/main/java/org/eclipse/milo/opcua/sdk/client/api/config/OpcUaClientConfig.java
@@ -13,7 +13,6 @@ package org.eclipse.milo.opcua.sdk.client.api.config;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-import org.eclipse.milo.opcua.binaryschema.parser.BsdParser;
 import org.eclipse.milo.opcua.sdk.client.api.identity.IdentityProvider;
 import org.eclipse.milo.opcua.stack.client.UaStackClientConfig;
 import org.eclipse.milo.opcua.stack.core.types.builtin.LocalizedText;
@@ -67,11 +66,6 @@ public interface OpcUaClientConfig extends UaStackClientConfig {
      * @return an {@link IdentityProvider} to use when activating a session.
      */
     IdentityProvider getIdentityProvider();
-
-    /**
-     * @return the {@link BsdParser} implementation used to serialize DataTypes defined by the server.
-     */
-    BsdParser getBsdParser();
 
     /**
      * @return the number of consecutive keep-alive request failures allowed before a connection is determined to be in
@@ -132,7 +126,6 @@ public interface OpcUaClientConfig extends UaStackClientConfig {
         builder.setMaxResponseMessageSize(config.getMaxResponseMessageSize());
         builder.setMaxPendingPublishRequests(config.getMaxPendingPublishRequests());
         builder.setIdentityProvider(config.getIdentityProvider());
-        builder.setBsdParser(config.getBsdParser());
         builder.setKeepAliveFailuresAllowed(config.getKeepAliveFailuresAllowed());
         builder.setKeepAliveInterval(config.getKeepAliveInterval());
         builder.setKeepAliveTimeout(config.getKeepAliveTimeout());
