+++ opc-ua-sdk/sdk-server/pom.xml
@@ -27,7 +27,7 @@
 
         <dependency>
             <groupId>org.eclipse.milo</groupId>
-            <artifactId>bsd-parser-core</artifactId>
+            <artifactId>bsd-generator</artifactId>
             <version>${project.version}</version>
         </dependency>
