+++ opc-ua-stack/bsd-core/src/main/resources/OPCBinarySchema.xsd
