--- opc-ua-stack/bsd-parser/bsd-parser-core/src/test/resources/dictionaries/BsdParserTest.bsd.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0"?>
-<opc:TypeDictionary DefaultByteOrder="LittleEndian"
-                    TargetNamespace="https://github.com/eclipse/milo"
-                    xmlns:opc="http://opcfoundation.org/BinarySchema/"
-                    xmlns:ua="http://opcfoundation.org/UA/"
-                    xmlns:tns="https://github.com/eclipse/milo">
-
-    <opc:StructuredType Name="Foo">
-        <opc:Field Name="Bar" TypeName="opc:Int32"/>
-        <opc:Field Name="Baz" TypeName="opc:String"/>
-    </opc:StructuredType>
-
-    <opc:StructuredType Name="Optionals">
-        <opc:Field Name="OptionalInt32Specified" TypeName="opc:Bit"/>
-        <opc:Field Name="OptionalStringSpecified" TypeName="opc:Bit"/>
-        <opc:Field Name="Reserved1" TypeName="opc:Bit" Length="30"/>
-
-        <opc:Field Name="OptionalInt32" TypeName="opc:Int32" SwitchField="OptionalInt32Specified"/>
-        <opc:Field Name="OptionalString" TypeName="opc:String" SwitchField="OptionalStringSpecified"/>
-    </opc:StructuredType>
-
-    <opc:StructuredType Name="ArrayContainer">
-        <opc:Field Name="IntArrayLen" TypeName="opc:Int32"/>
-        <opc:Field Name="IntArray" TypeName="opc:Int32" LengthField="IntArrayLen"/>
-
-        <opc:Field Name="BitField" TypeName="opc:Bit" Length="8"/>
-
-        <opc:Field Name="StringArrayLen" TypeName="opc:Int32"/>
-        <opc:Field Name="StringArray" TypeName="opc:String" LengthField="StringArrayLen"/>
-    </opc:StructuredType>
-
-    <opc:StructuredType Name="Bar">
-        <opc:Field Name="Foo" TypeName="tns:Foo"/>
-        <opc:Field Name="Int" TypeName="opc:Int32"/>
-        <opc:Field Name="Str" TypeName="opc:String"/>
-    </opc:StructuredType>
-
-    <opc:StructuredType Name="ProfilePointStruct">
-        <opc:Field TypeName="opc:Bit" Name="rangeXSpecified"/>
-        <opc:Field TypeName="opc:Bit" Name="rangeYSpecified"/>
-        <opc:Field Length="30" TypeName="opc:Bit" Name="Reserved1"/>
-        <opc:Field TypeName="opc:Double" Name="x"/>
-        <opc:Field TypeName="opc:Double" Name="y"/>
-        <opc:Field SwitchField="rangeXSpecified" TypeName="ua:Range" Name="rangeX"/>
-        <opc:Field SwitchField="rangeYSpecified" TypeName="ua:Range" Name="rangeY"/>
-    </opc:StructuredType>
-
-</opc:TypeDictionary>
