+++ opc-ua-stack/bsd-parser/src/main/java/org/eclipse/milo/opcua/binaryschema/parser/DictionaryDescription.java
