+++ opc-ua-stack/bsd-parser/src/test/java/org/eclipse/milo/opcua/binaryschema/GenericBsdParserTest.java
